- Moved Consumer and Publisher into separate folders - Created IConsumer and IPublisher for injection purposes
