+++ GeorgiaTechLib/Webshop.Tools.Messaging/IConsumer.cs
@@ -1,0 +1,7 @@
+﻿namespace Webshop.Tools.Messaging
+{
+    public interface IConsumer<T>
+    {
+        Task Connect(bool durable = false, bool exclusive = false, bool autoDelete = false);
+    }
+}
