+++ GeorgiaTechLib/Webshop.Tools.Messaging/IProducer.cs
@@ -1,0 +1,8 @@
+﻿namespace Webshop.Tools.Messaging
+{
+    public interface IProducer
+    {
+        Task SendMessage<T>(Message<T> message);
+        Task Connect(bool durable = false, bool exclusive = false, bool autoDelete = false);
+    }
+}
